requirements.txt
fastapi
python==3.11
python-dotenv
requests
